fix: close runner up sockets in the event there are multiple winners (#143)
The first attempt to fix this was to use the cpython staggered
race updates in #142 but there is still a race there where there
can be multiple winners. Instead we now accept that we will not
be able to cancel all coros in time and there will always be
a risk of multiple winners. We store all sockets in a set
that were not already cleaned up and we close all but the
first winner after the staggered race finishes.